requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
pillow-heif==0.7.2
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
